Rename and fix build break
