+++ packages/component/src/ConnectivityStatus/Connected.js
@@ -4,7 +4,9 @@ import ScreenReaderText from '../ScreenReaderText';
 import useLocalizer from '../hooks/useLocalizer';
 
 const ConnectivityStatusConnected = () => {
-  return <ScreenReaderText text={useLocalizer()('CONNECTIVITY_STATUS_ALT', useLocalize('CONNECTIVITY_STATUS_ALT_CONNECTED'))} />;
+  const localize = useLocalizer();
+
+  return <ScreenReaderText text={localize('CONNECTIVITY_STATUS_ALT', localize('CONNECTIVITY_STATUS_ALT_CONNECTED'))} />;
 };
 
 export default ConnectivityStatusConnected;
