+++ packages/component/src/Localization/en-US.json
@@ -10,26 +10,26 @@
   "CONNECTIVITY_STATUS_ALT_RECONNECTING": "Network interruption occurred. Reconnecting…",
   "CONNECTIVITY_STATUS_ALT_RENDER_ERROR": "Render error. Please check the console or contact the bot developer.",
 
-  "NOTIFICATION_ACCORDION_TWO": "$1 Notifications: Click here to see details",
-  "_NOTIFICATION_ACCORDION_TWO.comment": "$1 is the number of notifications. This is for plural rule of \"two\".",
-  "NOTIFICATION_ACCORDION_FEW": "$1 Notifications: Click here to see details",
-  "_NOTIFICATION_ACCORDION_FEW.comment": "$1 is the number of notifications. This is for plural rule of \"few\".",
-  "NOTIFICATION_ACCORDION_MANY": "$1 Notifications: Click here to see details",
-  "_NOTIFICATION_ACCORDION_MANY.comment": "$1 is the number of notifications. This is for plural rule of \"many\".",
-  "NOTIFICATION_ACCORDION_OTHER": "$1 Notifications: Click here to see details",
-  "_NOTIFICATION_ACCORDION_OTHER.comment": "$1 is the number of notifications. This is for plural rule of \"other\".",
+  "TOAST_ACCORDION_TWO": "$1 Notifications: Click here to see details",
+  "_TOAST_ACCORDION_TWO.comment": "$1 is the number of notifications. This is for plural rule of \"two\".",
+  "TOAST_ACCORDION_FEW": "$1 Notifications: Click here to see details",
+  "_TOAST_ACCORDION_FEW.comment": "$1 is the number of notifications. This is for plural rule of \"few\".",
+  "TOAST_ACCORDION_MANY": "$1 Notifications: Click here to see details",
+  "_TOAST_ACCORDION_MANY.comment": "$1 is the number of notifications. This is for plural rule of \"many\".",
+  "TOAST_ACCORDION_OTHER": "$1 Notifications: Click here to see details",
+  "_TOAST_ACCORDION_OTHER.comment": "$1 is the number of notifications. This is for plural rule of \"other\".",
 
-  "NOTIFICATION_ALT_ERROR": "Error: $1",
-  "_NOTIFICATION_ALT_ERROR.comment": "This is for screen reader. $1 is user-generated content, the actual notification message.",
-  "NOTIFICATION_ALT_INFO": "Information: $1",
-  "_NOTIFICATION_ALT_INFO.comment": "This is for screen reader. $1 is user-generated content, the actual notification message.",
-  "NOTIFICATION_ALT_SUCCESS": "Success: $1",
-  "_NOTIFICATION_ALT_SUCCESS.comment": "This is for screen reader. $1 is user-generated content, the actual notification message.",
-  "NOTIFICATION_ALT_WARN": "Warning: $1",
-  "_NOTIFICATION_ALT_WARN.comment": "This is for screen reader. $1 is user-generated content, the actual notification message.",
-  "NOTIFICATION_DISMISS_BUTTON": "Dismiss",
-  "NOTIFICATION_TITLE_ALT": "Notification",
-  "_NOTIFICATION_TITLE_ALT.comment": "This is for screen reader. The label for individiual notification.",
+  "TOAST_ALT_ERROR": "Error",
+  "_TOAST_ALT_ERROR.comment": "This is for screen reader.",
+  "TOAST_ALT_INFO": "Information",
+  "_TOAST_ALT_INFO.comment": "This is for screen reader.",
+  "TOAST_ALT_SUCCESS": "Success",
+  "_TOAST_ALT_SUCCESS.comment": "This is for screen reader.",
+  "TOAST_ALT_WARN": "Warning",
+  "_TOAST_ALT_WARN.comment": "This is for screen reader.",
+  "TOAST_DISMISS_BUTTON": "Dismiss",
+  "TOAST_TITLE_ALT": "Notification",
+  "_TOAST_TITLE_ALT.comment": "This is for screen reader. The label for individiual notification.",
 
   "ACTIVITY_BOT_SAID": "Bot $1 said, $2. Sent at $3.",
   "_ACTIVITY_BOT_SAID.comment": "$1 is the initials for the bot (e.g. \"JD\"), and $2 is user-generated content, the actual message.",
