+++ packages/component/src/Localization/yue.json
@@ -10,26 +10,26 @@
   "CONNECTIVITY_STATUS_ALT_RECONNECTING": "網絡暫時斷咗，試緊再駁返…",
   "CONNECTIVITY_STATUS_ALT_RENDER_ERROR": "Render 出事，請睇下 console 或者同 bot 開發人員聯絡。",
 
-  "NOTIFICATION_ACCORDION_TWO": "$1 通知：襟呢度睇詳情",
-  "_NOTIFICATION_ACCORDION_TWO.comment": "$1 is the number of notifications. This is for plural rule of \"two\".",
-  "NOTIFICATION_ACCORDION_FEW": "$1 通知：襟呢度睇詳情",
-  "_NOTIFICATION_ACCORDION_FEW.comment": "$1 is the number of notifications. This is for plural rule of \"few\".",
-  "NOTIFICATION_ACCORDION_MANY": "$1 通知：襟呢度睇詳情",
-  "_NOTIFICATION_ACCORDION_MANY.comment": "$1 is the number of notifications. This is for plural rule of \"many\".",
-  "NOTIFICATION_ACCORDION_OTHER": "$1 通知：襟呢度睇詳情",
-  "_NOTIFICATION_ACCORDION_OTHER.comment": "$1 is the number of notifications. This is for plural rule of \"other\".",
+  "TOAST_ACCORDION_TWO": "$1 通知：襟呢度睇詳情",
+  "_TOAST_ACCORDION_TWO.comment": "$1 is the number of notifications. This is for plural rule of \"two\".",
+  "TOAST_ACCORDION_FEW": "$1 通知：襟呢度睇詳情",
+  "_TOAST_ACCORDION_FEW.comment": "$1 is the number of notifications. This is for plural rule of \"few\".",
+  "TOAST_ACCORDION_MANY": "$1 通知：襟呢度睇詳情",
+  "_TOAST_ACCORDION_MANY.comment": "$1 is the number of notifications. This is for plural rule of \"many\".",
+  "TOAST_ACCORDION_OTHER": "$1 通知：襟呢度睇詳情",
+  "_TOAST_ACCORDION_OTHER.comment": "$1 is the number of notifications. This is for plural rule of \"other\".",
 
-  "NOTIFICATION_ALT_ERROR": "錯誤：$1",
-  "_NOTIFICATION_ALT_ERROR.comment": "This is for screen reader. $1 is user-generated content, the actual notification message.",
-  "NOTIFICATION_ALT_INFO": "資訊：$1",
-  "_NOTIFICATION_ALT_INFO.comment": "This is for screen reader. $1 is user-generated content, the actual notification message.",
-  "NOTIFICATION_ALT_SUCCESS": "成功：$1",
-  "_NOTIFICATION_ALT_SUCCESS.comment": "This is for screen reader. $1 is user-generated content, the actual notification message.",
-  "NOTIFICATION_ALT_WARN": "警告：$1",
-  "_NOTIFICATION_ALT_WARN.comment": "This is for screen reader. $1 is user-generated content, the actual notification message.",
-  "NOTIFICATION_DISMISS_BUTTON": "閂咗佢",
-  "NOTIFICATION_TITLE_ALT": "通知",
-  "_NOTIFICATION_TITLE_ALT.comment": "This is for screen reader. The label for individiual notification.",
+  "TOAST_ALT_ERROR": "錯誤：$1",
+  "_TOAST_ALT_ERROR.comment": "This is for screen reader. $1 is user-generated content, the actual notification message.",
+  "TOAST_ALT_INFO": "資訊：$1",
+  "_TOAST_ALT_INFO.comment": "This is for screen reader. $1 is user-generated content, the actual notification message.",
+  "TOAST_ALT_SUCCESS": "成功：$1",
+  "_TOAST_ALT_SUCCESS.comment": "This is for screen reader. $1 is user-generated content, the actual notification message.",
+  "TOAST_ALT_WARN": "警告：$1",
+  "_TOAST_ALT_WARN.comment": "This is for screen reader. $1 is user-generated content, the actual notification message.",
+  "TOAST_DISMISS_BUTTON": "閂咗佢",
+  "TOAST_TITLE_ALT": "通知",
+  "_TOAST_TITLE_ALT.comment": "This is for screen reader. The label for individiual notification.",
 
   "ACTIVITY_BOT_SAID": "Bot $1 話：$2。送出嘅時間：$3",
   "_ACTIVITY_BOT_SAID.comment": "$1 is the initials for the bot (e.g. \"JD\"), and $2 is user-generated content, the actual message.",
